Move jwt into auth directory
